Move backup directory
When using Python installed from the Microsoft Store (i.e. the UWP app
version), the %LOCALAPPDATA%/shroudstone directory is virtualized and
ends up somewhere obscure, so it's a bad place to put the backups
because we can't tell the user where to look! Weirdly, we seem to have
free reign over the actual %LOCALAPPDATA%/Stormgate directory, so all's
good to put it there I guess?