Fix error handling some maps
Custom 2v2 maps were not in the list and were defaulting to 1v1,
resulting in parsing errors.

IsleOfDread was listed as a 3-player map when it is in fact 2-player.

I've gotten rid of the assumption that any unknown map is 2-player,
I think it's better to immediately raise an error than to crash later or
to produce incorrect results.